获取开关类触点

+++ Circuit3D/engine/src/main/java/com/cas/circuit/element/ResistorElm.java
@@ -12,7 +12,7 @@ public class ResistorElm extends CircuitElm {
 	public ResistorElm() {
 	}
 
-	public ResistorElm(int r) {
+	public ResistorElm(double r) {
 		resistance = r;
 	}
